+++ flask_unchained/bundles/mail/commands.py
@@ -1,11 +1,9 @@
-import click
-
-from flask.cli import with_appcontext
+from flask_unchained.cli import cli, click
 
 from .extensions import mail as mail_ext
 
 
-@click.group()
+@cli.group()
 def mail():
     """Mail commands"""
 
@@ -14,6 +12,5 @@ def mail():
 @click.option('--subject', help='Message subject',
               default='Hello world from Flask Mail')
 @click.option('--to', help='Message recipient')
-@with_appcontext
 def send_test_email(subject, to):
     mail_ext.send(subject, to, template='email/__test_email__.html')